# artifacts/api-server/src/routes/openai/index.ts
import { Router, type IRouter } from "express";
import { eq, and } from "drizzle-orm";
import { db, conversations, messages } from "@workspace/db";
import { recordQuestionForUser } from "../streaks";
import {
  openai,
  CHAT_MODEL,
  ACTIVE_PROVIDER,
  visionOpenai,
  VISION_MODEL,
  ACTIVE_VISION_PROVIDER,
  type ChatCompletionMessageParam,
  type ChatCompletionContentPart,
} from "@workspace/integrations-openai-ai-server";
import { requireAuth } from "../../lib/require-auth";
import { requireQuota } from "../../lib/require-quota";
import { resolveAttachments, hasImages, buildAttachmentPromptText } from "../../lib/attachments";
import { recordAiUsage } from "../../lib/aiCost";
import {
  CreateOpenaiConversationBody,
  GetOpenaiConversationParams,
  DeleteOpenaiConversationParams,
  ListOpenaiMessagesParams,
  SendOpenaiMessageParams,
  SendOpenaiMessageBody,
} from "@workspace/api-zod";

const router: IRouter = Router();
router.use(requireAuth);

// ─── Cambridge Verification Pipeline ────────────────────────────────────────

interface VerificationResult {
  syllabusRef: string | null;
  markSchemePoints: string[];
  confidence: "high" | "medium" | "low";
  examinerNote: string | null;
}

async function verifyQuestion(
  question: string,
  subject: string | null,
  level: string | null,
): Promise<VerificationResult> {
  const subjectCtx = [subject, level].filter(Boolean).join(" ") || "unknown subject";
  try {
    const response = await openai.chat.completions.create({
      model: CHAT_MODEL,
      max_tokens: 220,
      temperature: 0,
      messages: [
        {
          role: "user",
          content: `You are a Cambridge International Examinations expert. A student studying ${subjectCtx} has asked or is about to receive a response to:

"${question.slice(0, 400)}"

Respond with ONLY valid JSON — no markdown fences, no explanation:
{
  "syllabusRef": "exact syllabus code + learning objective (e.g. 'Physics 9702 LO 5.1.3') or null if you cannot confirm it precisely",
  "markSchemePoints": ["what examiners actually award marks for — max 4 bullet points — use Cambridge mark-scheme language"],
  "confidence": "high if you are certain of the exact mark scheme, medium if mostly certain, low if genuinely uncertain",
  "examinerNote": "relevant warning from Cambridge examiner reports (e.g. common errors, misconceptions examiners flag) or null"
}

CRITICAL: Do NOT invent syllabus references. If uncertain, set syllabusRef to null and confidence to low. Accuracy over confidence.`,
        },
      ],
    });
    const raw = (response.choices[0]?.message?.content ?? "{}").trim();
    const parsed = JSON.parse(raw.replace(/^```json\n?/, "").replace(/\n?```$/, ""));
    return {
      syllabusRef: typeof parsed.syllabusRef === "string" ? parsed.syllabusRef : null,
      markSchemePoints: Array.isArray(parsed.markSchemePoints) ? parsed.markSchemePoints.slice(0, 4) : [],
      confidence: (["high", "medium", "low"] as const).includes(parsed.confidence) ? parsed.confidence : "medium",
      examinerNote: typeof parsed.examinerNote === "string" ? parsed.examinerNote : null,
    };
  } catch {
    return { syllabusRef: null, markSchemePoints: [], confidence: "medium", examinerNote: null };
  }
}

function buildVerificationContext(v: VerificationResult): string {
  const parts: string[] = [
    "════════════════════════════════",
    "VERIFICATION CONTEXT (checked before answering)",
    "════════════════════════════════",
  ];
  if (v.syllabusRef) parts.push(`Syllabus reference: ${v.syllabusRef}`);
  if (v.markSchemePoints.length > 0) {
    parts.push("Mark scheme key points:");
    v.markSchemePoints.forEach((p) => parts.push(`  • ${p}`));
  }
  parts.push(`Confidence in mark scheme: ${v.confidence.toUpperCase()}`);
  if (v.examinerNote) parts.push(`Examiner report note: ${v.examinerNote}`);
  parts.push(
    v.confidence === "low"
      ? "INSTRUCTION: Explicitly tell the student you are not certain of the exact mark scheme for this. Advise them to verify with official Cambridge past paper mark schemes at cambridgeinternational.org. Still give your best answer but be transparent about uncertainty."
      : "INSTRUCTION: Use these mark scheme points to frame your answer. Align your response to what Cambridge examiners actually award marks for.",
  );
  return parts.join("\n");
}

// ────────────────────────────────────────────────────────────────────────────

function userFilter(req: Parameters<Parameters<typeof router.get>[1]>[0]) {
  return eq(conversations.userId, req.user!.id);
}

const CAMBRIDGE_SYSTEM_PROMPT = `You are CamAI — an elite AI study tutor built exclusively for Cambridge International Examinations (O Level and A Level). You are not a generic AI. You know the Cambridge system inside out.

════════════════════════════════
CAMBRIDGE O LEVEL SUBJECTS YOU COVER
════════════════════════════════
• Mathematics (4024) — algebra, geometry, trigonometry, statistics, number
• Additional Mathematics (4037) — functions, calculus, binomial, vectors, logarithms
• Physics (5054) — mechanics, waves, electricity, magnetism, nuclear, thermal
• Chemistry (5070) — atomic structure, bonding, organic chemistry, redox, energetics
• Biology (5090) — cell biology, genetics, ecology, physiology, reproduction
• Combined Science (5129) — core concepts across Physics, Chemistry, Biology
• English Language (1123) — directed writing, composition, comprehension, summary
• English Literature (2010) — poetry, prose, drama analysis
• History (2147) — source-based, essay skills, key events from 1900-2000
• Geography (2217) — physical and human geography, map skills, fieldwork
• Economics (2281) — micro/macro, supply & demand, market structures, development
• Commerce (7100) — trade, banking, insurance, transport, communications
• Accounting (7707) — double-entry bookkeeping, financial statements, ratios, ledger accounts
• Business Studies (7115) — business organisation, HR, marketing, finance, operations
• Computer Science (2210) — algorithms, programming, data representation, networks, security
• Information & Communication Technology (2210) — applications, systems, digital literacy
• Design & Technology (6043) — design process, materials, structures, mechanisms, manufacturing
• Food & Nutrition (6065) — nutrients, food science, diet, health, food safety
• Art & Design (6010) — visual elements, practical work, critical analysis
• Islamiyat (2058) — Quran, Hadith, Islamic beliefs, practices, history
• Pakistan Studies (2059) — Pakistan's geography, history, culture, government
• Sociology (2251) — socialisation, institutions, stratification, methods
• Environmental Management (5014) — ecosystems, resources, sustainability, pollution
• Urdu (3247) — reading, writing, comprehension, composition
• Bengali (3204) — reading, writing, comprehension
• Arabic (3180) — language skills, comprehension
• Chinese (3247) — reading, writing, characters, grammar
• Malay (3820) — language skills, composition
• French (3015) — language skills, grammar, composition
• Tamil — language skills, reading, writing

════════════════════════════════
CAMBRIDGE A LEVEL SUBJECTS YOU COVER
════════════════════════════════
• Mathematics (9709) — pure maths (P1–P3), mechanics (M1–M2), statistics (S1–S2)
• Further Mathematics (9231) — FP1, FP2, mechanics, statistics, discrete
• Physics (9702) — AS and A2 full syllabus, practical skills, paper 5
• Chemistry (9701) — AS and A2, organic/inorganic/physical, Paper 5
• Biology (9700) — AS and A2, genetics, evolution, ecology, physiology
• English Language (9093) — language analysis, writing for different purposes
• English Literature (9695) — critical essays, unseen texts, comparative analysis
• History (9489) — 19th–21st century topics, interpretations, essay writing
• Geography (9696) — physical systems, human environments, global issues
• Economics (9708) — microeconomics, macroeconomics, evaluation essays
• Accounting (9706) — financial statements, management accounting, analysis
• Business (9609) — strategic management, finance, operations, marketing
• Computer Science (9618) — algorithms, OOP, data structures, networking, AI
• Information Technology (9626) — systems, database, networks, project
• Design & Technology (9705) — advanced design process, materials, systems
• Law (9084) — contract, tort, criminal law, legal reasoning
• Psychology (9990) — research methods, cognitive, social, biological, abnormal
• Sociology (9699) — sociological theory, stratification, institutions, research
• Thinking Skills (9694) — critical thinking, problem solving
• Art & Design (9479) — coursework, critical studies, studio practice
• General Paper (8021) — essays, comprehension, current affairs, argument
• Urdu (9686), Arabic (9680), Chinese Language (9715), Malay (9679), French (9716), Tamil

════════════════════════════════
HOW YOU HELP — WHAT SETS YOU APART
════════════════════════════════
1. CAMBRIDGE COMMAND WORDS — you know exactly what each means:
   • "Define" → give precise meaning, no examples needed
   • "State" → brief factual answer, no explanation
   • "Describe" → give features/characteristics
   • "Explain" → give reason/mechanism (use "because", "therefore", "this causes")
   • "Discuss" → balanced argument with evidence on both sides
   • "Evaluate" → make a judgement, weigh evidence, reach a conclusion
   • "Calculate" → show all working, include units
   • "Sketch" → labelled diagram, no need for accuracy
   • "Suggest" → apply knowledge to unfamiliar situation

2. MARK SCHEME THINKING — you frame answers the way examiners mark them:
   • For 2-mark answers: point + explanation
   • For 4-mark answers: two developed points
   • For 6-mark+ essays: clear structure, PEE (Point, Evidence, Explain), conclusion

3. SUBJECT-SPECIFIC EXPERTISE:
   • Maths/Add Maths/Further Maths: show every step, state theorems, use correct notation
   • Sciences: use precise scientific vocabulary, show equation → substitution → answer with units
   • Accounting: balance every account, show T-accounts, state the accounting principle used
   • Business/Economics: always use a diagram where relevant, evaluate both sides
   • Languages: give model sentences, grammar rules, vocabulary in context
   • Design & Technology: reference materials, processes, sustainability, ergonomics
   • Food & Nutrition: link nutrients to functions and deficiency diseases
   • Islamiyat: quote Quranic references and Hadith accurately
   • Pakistan Studies: use precise dates, names of leaders, and geographical facts

4. STUDY STRATEGIES tailored to Cambridge:
   • Topic-by-topic breakdown using the official syllabus learning outcomes
   • Past paper technique and time management (per paper type)
   • Active recall through practice questions you generate
   • Spaced repetition suggestions for key definitions and formulas

════════════════════════════════
QUESTION GENERATION — CAMBRIDGE STYLE (MANDATORY)
════════════════════════════════
Whenever you give a student ANY practice question, exam question, or revision question — invented or from memory — follow ALL of these rules without exception:

1. ALWAYS show the mark allocation in square brackets immediately after the question:
   "State one function of the mitochondria. [1]"
   "Explain why increasing temperature increases the rate of a reaction. [3]"
   "Discuss the advantages and disadvantages of monopoly. [8]"
   The mark allocation tells the student exactly how much depth is expected. Never omit it.

2. MATCH the command word to the mark allocation every single time:
   • [1]     → State / Name / Identify / Give / Write down / Define
   • [2]     → State and explain / Give one reason with explanation / Describe briefly
   • [3–4]   → Explain / Describe in detail / Calculate (show full working) / Compare
   • [5–6]   → Discuss / Analyse / Examine / Assess
   • [7–10]  → Evaluate / To what extent / Essay-style (intro + developed body + conclusion)
   Using the wrong command word for the mark value is a mismatch — never do it.

3. FORMAT questions by paper type:
   • Structured papers (Paper 2, 4): use sub-parts (a)(i), (a)(ii), (b)(i) etc. with marks on each part
   • MCQ (Paper 1): four options A–D, exactly one correct, label the paper
   • Data response / case study (Economics, Business): 3–5 line scenario stem, then numbered questions
   • Sciences practical (Paper 3/5): present data/graph/table description, then question
   • Essay papers (History, Literature, GP): clear rubric and mark band guidance

4. INVENT with full Cambridge authenticity:
   • Use real-world contexts Cambridge favours (e.g. Physics: roller coasters, power stations; Chemistry: industrial Haber process; Economics: real countries and policies; History: real events and sources)
   • Match the difficulty band of the actual paper — not easier, not harder
   • Use precise subject vocabulary exactly as it appears in Cambridge mark schemes
   • Sciences: always include units, significant figures guidance, and state the formula required for calculation questions

5. AFTER a student answers, always mark it Cambridge-style:
   Show a mini mark scheme:
   ✓ [credited point] — mark awarded
   ✗ [missed point] — what was expected
   Total: X / Y marks
   Then explain what a full-mark answer would look like.

════════════════════════════════
ALWAYS
════════════════════════════════
- Use Cambridge syllabus language and mark scheme style
- For maths and science: equation → substitute → solve → units
- For essays: clear intro, developed body, evaluative conclusion
- Tell students which paper/component a question type comes from
- Be warm and encouraging — exam stress is real
- When unsure, say so honestly rather than guess
- Respond in the same language the student uses (English, Urdu, Arabic, Malay, etc.)`;


router.get("/openai/conversations", async (req, res): Promise<void> => {
  const convs = await db
    .select()
    .from(conversations)
    .where(userFilter(req))
    .orderBy(conversations.createdAt);
  res.json(convs);
});

router.post("/openai/conversations", async (req, res): Promise<void> => {
  const parsed = CreateOpenaiConversationBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [conv] = await db
    .insert(conversations)
    .values({
      title: parsed.data.title,
      subject: parsed.data.subject ?? null,
      level: parsed.data.level ?? null,
       userId: req.user!.id,
    })
    .returning();
  res.status(201).json(conv);
});

router.get("/openai/conversations/:id", async (req, res): Promise<void> => {
  const params = GetOpenaiConversationParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  const [conv] = await db
    .select()
    .from(conversations)
    .where(and(eq(conversations.id, params.data.id), userFilter(req)));
  if (!conv) {
    res.status(404).json({ error: "Conversation not found" });
    return;
  }
  const msgs = await db
    .select()
    .from(messages)
    .where(eq(messages.conversationId, params.data.id))
    .orderBy(messages.createdAt);
  res.json({ ...conv, messages: msgs });
});

router.delete("/openai/conversations/:id", async (req, res): Promise<void> => {
  const params = DeleteOpenaiConversationParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  const [deleted] = await db
    .delete(conversations)
    .where(and(eq(conversations.id, params.data.id), userFilter(req)))
    .returning();
  if (!deleted) {
    res.status(404).json({ error: "Conversation not found" });
    return;
  }
  res.sendStatus(204);
});

router.get("/openai/conversations/:id/messages", async (req, res): Promise<void> => {
  const params = ListOpenaiMessagesParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  // verify ownership before returning messages
  const [conv] = await db
    .select()
    .from(conversations)
    .where(and(eq(conversations.id, params.data.id), userFilter(req)));
  if (!conv) {
    res.status(404).json({ error: "Conversation not found" });
    return;
  }
  const msgs = await db
    .select()
    .from(messages)
    .where(eq(messages.conversationId, params.data.id))
    .orderBy(messages.createdAt);
  res.json(msgs);
});

router.post("/openai/conversations/:id/messages", requireQuota("tutorMessage"), async (req, res): Promise<void> => {
  const params = SendOpenaiMessageParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  const body = SendOpenaiMessageBody.safeParse(req.body);
  if (!body.success) {
    res.status(400).json({ error: body.error.message });
    return;
  }

  const [conv] = await db
    .select()
    .from(conversations)
    .where(and(eq(conversations.id, params.data.id), userFilter(req)));
  if (!conv) {
    res.status(404).json({ error: "Conversation not found" });
    return;
  }

  await db.insert(messages).values({
    conversationId: params.data.id,
    role: "user",
    content: body.data.content,
  });
  if (req.isAuthenticated()) {
    await recordQuestionForUser(req.user.id);
  }

  const history = await db
    .select()
    .from(messages)
    .where(eq(messages.conversationId, params.data.id))
    .orderBy(messages.createdAt);

  // Attachments are resolved fresh for this request only — the extracted
  // text/images are never persisted to the message row, so a follow-up
  // question later in the conversation won't still see them (a known
  // limitation of this first pass, not an oversight).
  const attachments = await resolveAttachments(req.user!.id, body.data.attachmentIds ?? []);

  // Phase 1: fast verification (runs before streaming starts)
  const verification = await verifyQuestion(body.data.content, conv.subject, conv.level);

  const subjectContext = conv.subject
    ? ` Focus on ${conv.subject} at ${conv.level ?? "Cambridge"} level.`
    : "";

  const verificationContext = buildVerificationContext(verification);

  const lastMessageIndex = history.length - 1;
  const chatMessages: ChatCompletionMessageParam[] = [
    { role: "system", content: CAMBRIDGE_SYSTEM_PROMPT + subjectContext + "\n\n" + verificationContext },
    ...history.map((m, i) => {
      if (i === lastMessageIndex && m.role === "user" && attachments.length > 0) {
        const parts: ChatCompletionContentPart[] = [
          { type: "text", text: buildAttachmentPromptText(m.content, attachments) },
        ];
        for (const a of attachments) {
          if (a.kind === "image") {
            parts.push({ type: "image_url", image_url: { url: a.imageDataUrl! } });
          }
        }
        return { role: "user" as const, content: parts };
      }
      return { role: m.role as "user" | "assistant", content: m.content };
    }),
  ];

  const useVision = hasImages(attachments);
  const client = useVision ? visionOpenai : openai;
  const model = useVision ? VISION_MODEL : CHAT_MODEL;

  res.setHeader("Content-Type", "text/event-stream");
  res.setHeader("Cache-Control", "no-cache");
  res.setHeader("Connection", "keep-alive");

  // Send verification metadata to frontend before content starts
  res.write(`data: ${JSON.stringify({ type: "verification", ...verification })}\n\n`);

  let fullResponse = "";
  const stream = await client.chat.completions.create({
    model,
    max_tokens: 8192,
    messages: chatMessages,
    stream: true,
    stream_options: { include_usage: true },
  });

  let usage: { prompt_tokens?: number; completion_tokens?: number } | null = null;
  for await (const chunk of stream) {
    const content = chunk.choices[0]?.delta?.content;
    if (content) {
      fullResponse += content;
      res.write(`data: ${JSON.stringify({ content })}\n\n`);
    }
    if (chunk.usage) usage = chunk.usage;
  }

  await recordAiUsage({
    userId: req.user!.id,
    feature: "tutorMessage",
    provider: useVision ? ACTIVE_VISION_PROVIDER : ACTIVE_PROVIDER,
    model,
    promptTokens: usage?.prompt_tokens ?? 0,
    completionTokens: usage?.completion_tokens ?? 0,
  });

  await db.insert(messages).values({
    conversationId: params.data.id,
    role: "assistant",
    content: fullResponse,
  });

  res.write(`data: ${JSON.stringify({ done: true })}\n\n`);
  res.end();
});

router.post("/openai/revision-stream", requireQuota("tutorMessage"), async (req, res): Promise<void> => {
  const { messages: msgs, system, subject, level } = req.body as {
    messages: { role: string; content: string }[];
    system?: string;
    subject?: string;
    level?: string;
  };

  if (!Array.isArray(msgs)) {
    res.status(400).json({ error: "messages must be an array" });
    return;
  }

  // Phase 1: verify the latest user message
  const lastUserMsg = [...msgs].reverse().find((m) => m.role === "user");
  const verification = lastUserMsg
    ? await verifyQuestion(lastUserMsg.content, subject ?? null, level ?? null)
    : { syllabusRef: null, markSchemePoints: [], confidence: "high" as const, examinerNote: null };

  const verificationContext = buildVerificationContext(verification);

  const chatMessages: { role: "system" | "user" | "assistant"; content: string }[] = [
    { role: "system", content: (system ?? CAMBRIDGE_SYSTEM_PROMPT) + "\n\n" + verificationContext },
    ...msgs.map((m) => ({
      role: m.role as "user" | "assistant",
      content: m.content,
    })),
  ];

  res.setHeader("Content-Type", "text/event-stream");
  res.setHeader("Cache-Control", "no-cache");
  res.setHeader("Connection", "keep-alive");

  // Send verification metadata before content stream
  res.write(`data: ${JSON.stringify({ type: "verification", ...verification })}\n\n`);

  const stream = await openai.chat.completions.create({
    model: CHAT_MODEL,
    max_tokens: 8192,
    messages: chatMessages,
    stream: true,
    stream_options: { include_usage: true },
  });

  let usage: { prompt_tokens?: number; completion_tokens?: number } | null = null;
  for await (const chunk of stream) {
    const content = chunk.choices[0]?.delta?.content;
    if (content) {
      res.write(`data: ${JSON.stringify({ content })}\n\n`);
    }
    if (chunk.usage) usage = chunk.usage;
  }

  await recordAiUsage({
    userId: req.user!.id,
    feature: "tutorMessage",
    provider: ACTIVE_PROVIDER,
    model: CHAT_MODEL,
    promptTokens: usage?.prompt_tokens ?? 0,
    completionTokens: usage?.completion_tokens ?? 0,
  });

  res.write(`data: ${JSON.stringify({ done: true })}\n\n`);
  res.end();
});

export default router;